requirements.txt
asgiref==3.7.2
distro==1.8.0
Django==4.2.1
djangorestframework==3.14.0
numpy==1.24.3
pandas==2.0.2
pdfkit==1.0.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tabula-py==2.7.0
tzdata==2023.3
